fix: break circular dependency in ces-contracts RpcErrorSchema import
rpc.ts imported RpcErrorSchema from index.ts, but index.ts re-exports from
rpc.ts, creating a circular reference that caused 'Cannot access before
initialization' at runtime. Extract RpcErrorSchema into error.ts so both
index.ts and rpc.ts can import it without cycles.
